+++ account_man.py
@@ -3,4 +3,8 @@ from account import Account
 
 class AccountManager(object):
 	def __init__(self):
-		
+		self.man = Manager() 
+		self.query = ""
+	def _table_clients(self):
+		columns = []
+		columns.append(self.)
